Move errorMsg function to shared folder
