+++ packages/react-strict-dom/src/shared/errorMsg.js
